Refactor delivery runner into dedicated module
