+++ backend/services/deliveries.py
@@ -1,8 +1,7 @@
-"""Delivery service for managing delivery jobs and execution."""
+"""Delivery service for managing delivery jobs and queue hand-off."""
 
 from __future__ import annotations
 
-import asyncio
 import json
 from datetime import datetime, timezone
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
@@ -16,7 +15,6 @@ from backend.models import DeliveryJob
 if TYPE_CHECKING:  # pragma: no cover - used for type checking only
     from .queue import QueueBackend
 
-_SUCCESS_STATUSES = {"ok", "completed", 200}
 _ACTIVE_STATUSES = {"pending", "running", "retrying"}
 
 
@@ -263,124 +261,3 @@ class DeliveryService:
             return json.loads(job.result)
         except json.JSONDecodeError:
             return None
-
-
-def process_delivery_job(
-    job_id: str,
-    retries_left: Optional[int] = None,
-    *,
-    raise_on_error: bool = False,
-) -> None:
-    """Process a delivery job synchronously."""
-
-    async def runner() -> None:
-        await _process_delivery_job_async(
-            job_id,
-            retries_left=retries_left,
-            raise_on_error=raise_on_error,
-        )
-
-    try:
-        loop = asyncio.get_running_loop()
-    except RuntimeError:
-        asyncio.run(runner())
-    else:  # pragma: no cover - defensive branch
-        task = loop.create_task(runner())
-        if raise_on_error:
-            raise RuntimeError(
-                "Cannot raise errors when scheduling delivery processing on an"
-                " active event loop",
-            )
-        return task
-
-
-async def _process_delivery_job_async(
-    job_id: str,
-    *,
-    retries_left: Optional[int] = None,
-    raise_on_error: bool = False,
-) -> None:
-    """Internal coroutine to process a delivery job."""
-
-    from backend.core.database import get_session_context
-
-    prompt: str
-    mode: str
-    params: Dict[str, Any]
-    error: Optional[Exception] = None
-    result_payload: Dict[str, Any] = {}
-    status: str = "failed"
-
-    with get_session_context() as session:
-        service = DeliveryService(session)
-        job = service.get_job(job_id)
-        if job is None:
-            return
-
-        service.update_job_status(job_id, "running")
-        prompt = job.prompt
-        mode = job.mode
-        params = service.get_job_params(job)
-
-    try:
-        result_payload = await _execute_delivery_backend(prompt, mode, params)
-        status = "succeeded" if _is_successful_result(result_payload) else "failed"
-    except Exception as exc:  # pragma: no cover - defensive branch
-        error = exc
-        result_payload = {"error": str(exc)}
-        if retries_left is not None and retries_left > 0:
-            result_payload["retries_left"] = retries_left
-            status = "retrying"
-        else:
-            status = "failed"
-
-    with get_session_context() as session:
-        service = DeliveryService(session)
-        service.update_job_status(job_id, status, result_payload)
-
-    if error is not None and raise_on_error:
-        raise error
-
-
-async def _execute_delivery_backend(
-    prompt: str,
-    mode: str,
-    params: Dict[str, Any],
-) -> Dict[str, Any]:
-    """Execute the delivery backend for a job and return the result."""
-
-    if mode == "http":
-        from backend.delivery import get_delivery_backend
-
-        backend = get_delivery_backend("http")
-        return await backend.deliver(prompt, params)
-    if mode == "cli":
-        from backend.delivery import get_delivery_backend
-
-        backend = get_delivery_backend("cli")
-        return await backend.deliver(prompt, params)
-    if mode == "sdnext":
-        from backend.delivery import get_generation_backend
-        from backend.schemas import SDNextGenerationParams
-
-        backend = get_generation_backend("sdnext")
-
-        gen_params_dict = params.get("generation_params", {}).copy()
-        gen_params_dict["prompt"] = prompt
-
-        gen_params = SDNextGenerationParams(**gen_params_dict)
-        full_params = {
-            "generation_params": gen_params.model_dump(),
-            "mode": params.get("mode", "immediate"),
-            "save_images": params.get("save_images", True),
-            "return_format": params.get("return_format", "base64"),
-        }
-        result_obj = await backend.generate_image(prompt, full_params)
-        return result_obj.model_dump()
-
-    return {"status": "error", "detail": f"unknown mode: {mode}"}
-
-
-def _is_successful_result(result: Dict[str, Any]) -> bool:
-    status = result.get("status")
-    return status in _SUCCESS_STATUSES
